Move the table creation into a dedicated module full of all the heredocs and executes
